linted

+++ yolo/ops/loss_utils.py
@@ -18,7 +18,7 @@ def _build_grid_points(lwidth, lheight, anchors, dtype):
   return x_y
 
 
-def _build_anchor_grid(width, height, anchors, dtype): #, num, dtype):
+def _build_anchor_grid(width, height, anchors, dtype):  #, num, dtype):
   with tf.name_scope('anchor_grid'):
     """ get the transformed anchor boxes for each dimention """
     num = tf.shape(anchors)[0]
@@ -49,7 +49,6 @@ class GridGenerator(object):
   def _extend_batch(self, grid, batch_size):
     return tf.tile(grid, [batch_size, 1, 1, 1, 1])
 
-
   def __call__(self, width, height, batch_size, dtype=None):
     if dtype is None:
       self.dtype = tf.keras.backend.floatx()
@@ -59,7 +58,8 @@ class GridGenerator(object):
     anchor_grid = _build_anchor_grid(
         width, height,
         tf.cast(self._anchors, self.dtype) /
-        tf.cast(self._scale_anchors * width, self.dtype), self.dtype) #self._num, self.dtype)
+        tf.cast(self._scale_anchors * width, self.dtype),
+        self.dtype)  #self._num, self.dtype)
     grid_points = self._extend_batch(grid_points, batch_size)
     anchor_grid = self._extend_batch(anchor_grid, batch_size)
     return tf.stop_gradient(grid_points), tf.stop_gradient(anchor_grid)
